Add dir argument to clean.sh

+++ clean.sh
@@ -1,3 +1,13 @@
 #!/bin/bash
-rm -r output/*
-rm slurm*
+
+# Args: [dir]
+
+DIR=$1
+
+if [ -z "$DIR" ] ; then
+  DIR="output"
+  rm slurm*
+fi
+
+rm -r "DIR"
+
